requirements.txt
pandas==2.3.3
numpy==2.1.3
scikit-learn==1.6.1
lightgbm==4.6.0
xgboost==3.0.5
catboost==1.2.10
joblib==1.5.3
fastapi==0.135.3
prometheus-client==0.21.1
prometheus-fastapi-instrumentator==6.1.0
httpx==0.28.1
uvicorn[standard]==0.44.0
pyarrow==21.0.0
kagglehub==0.3.13
mlflow==3.11.1
optuna==4.6.0
evidently==0.7.21
pyyaml==6.0.2
pytest==8.3.4
flake8==7.1.1
